# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Webingo, a bingo game application.",
        "title": "Webingo API",
        "contact": {},
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/boards": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new board with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Create a new board",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Board"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/boards/{boardID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a board by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Get a board by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Board"
                        }
                    },
                    "400": {
                        "description": "Invalid board ID",
                        "schema": {}
                    },
                    "404": {
                        "description": "Board not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates an existing board with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Update a board",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Board"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {}
                    },
                    "404": {
                        "description": "Board not found",
                        "schema": {}
                    },
                    "409": {
                        "description": "Data version mismatch",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a board by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Delete a board",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid board ID",
                        "schema": {}
                    },
                    "404": {
                        "description": "Board not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/boards/{boardID}/comments": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new comment on the specified board",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Create a comment on a board",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.BoardComment"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {}
                    },
                    "404": {
                        "description": "Board not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Board": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BoardComment"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "spaces": {
                    "description": "GORM tag, no json tag if Space has Board reference",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Space"
                    }
                },
                "tags": {
                    "description": "Tags for the board",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "version": {
                    "description": "For optimistic locking",
                    "type": "integer"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BoardVote"
                    }
                }
            }
        },
        "model.BoardCollection": {
            "type": "object",
            "properties": {
                "boards": {
                    "description": "Many-to-many relationship with Boards",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Board"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "description": "Optional description",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isPublic": {
                    "description": "Can others view this collection?",
                    "type": "boolean"
                },
                "name": {
                    "description": "Name of the collection (e.g., \"My Party Boards\")",
                    "type": "string"
                },
                "type": {
                    "description": "e.g., \"FAVORITES\", \"USER_CREATED\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.BoardCollectionType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "description": "User who owns the collection",
                    "type": "string"
                }
            }
        },
        "model.BoardCollectionType": {
            "type": "string",
            "enum": [
                "FAVORITES",
                "USER_CREATED"
            ],
            "x-enum-varnames": [
                "BoardCollectionTypeFavorites",
                "BoardCollectionTypeUserCreated"
            ]
        },
        "model.BoardComment": {
            "type": "object",
            "properties": {
                "boardId": {
                    "type": "string"
                },
                "content": {
                    "description": "Content of the comment",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.BoardVote": {
            "type": "object",
            "properties": {
                "board": {
                    "$ref": "#/definitions/model.Board"
                },
                "boardId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isUpvote": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.Cell": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "gameId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "posH": {
                    "type": "integer"
                },
                "posW": {
                    "type": "integer"
                },
                "spaceId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Game": {
            "type": "object",
            "properties": {
                "board": {
                    "$ref": "#/definitions/model.Board"
                },
                "boardId": {
                    "type": "string"
                },
                "cells": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Cell"
                    }
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GameComment"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "finished": {
                    "type": "boolean"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "playerMoves": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PlayerMove"
                    }
                },
                "private": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "description": "User who created the game, optional for anonymous games",
                    "type": "string"
                },
                "votes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GameVote"
                    }
                },
                "width": {
                    "type": "integer"
                },
                "youtubeVideoId": {
                    "type": "string"
                }
            }
        },
        "model.GameComment": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the comment",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "gameId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.GameVote": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "game": {
                    "$ref": "#/definitions/model.Game"
                },
                "gameId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isUpvote": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.Notification": {
            "type": "object",
            "properties": {
                "acknowledgedAt": {
                    "description": "When the user dismissed/acted on it",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isRead": {
                    "description": "Has the user read it?",
                    "type": "boolean"
                },
                "link": {
                    "description": "Optional URL to navigate to",
                    "type": "string"
                },
                "message": {
                    "description": "Display message",
                    "type": "string"
                },
                "relatedEntityId": {
                    "description": "Optional ID of related entity (e.g., BoardID, CommentID)",
                    "type": "string"
                },
                "relatedEntityType": {
                    "description": "Optional type of related entity (e.g., \"Board\", \"Comment\")",
                    "type": "string"
                },
                "type": {
                    "description": "Type of notification",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.NotificationType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "description": "User to whom the notification belongs",
                    "type": "string"
                }
            }
        },
        "model.NotificationType": {
            "type": "string",
            "enum": [
                "COMMENT",
                "FOLLOW",
                "VOTE",
                "SYSTEM",
                "UPDATE"
            ],
            "x-enum-comments": {
                "NotificationTypeUpdate": "e.g., board you favorited updated"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "e.g., board you favorited updated"
            ],
            "x-enum-varnames": [
                "NotificationTypeComment",
                "NotificationTypeFollow",
                "NotificationTypeVote",
                "NotificationTypeSystem",
                "NotificationTypeUpdate"
            ]
        },
        "model.PlayerMove": {
            "type": "object",
            "properties": {
                "cellH": {
                    "type": "integer"
                },
                "cellW": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "game": {
                    "$ref": "#/definitions/model.Game"
                },
                "gameId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Report": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "description": "Optional detailed description",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isResolved": {
                    "description": "Has the report been actioned by moderation?",
                    "type": "boolean"
                },
                "moderatorUser": {
                    "$ref": "#/definitions/model.User"
                },
                "moderatorUserId": {
                    "description": "ID of the moderator who handled it",
                    "type": "string"
                },
                "reason": {
                    "description": "Reason for the report",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.ReportReason"
                        }
                    ]
                },
                "reportedEntityId": {
                    "description": "ID of the entity being reported",
                    "type": "string"
                },
                "reportedEntityType": {
                    "description": "Type of the entity (e.g., \"BOARD\", \"COMMENT\", \"USER\")",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.ReportedEntityType"
                        }
                    ]
                },
                "reporterUser": {
                    "$ref": "#/definitions/model.User"
                },
                "reporterUserId": {
                    "description": "User who submitted the report",
                    "type": "string"
                },
                "resolutionNotes": {
                    "description": "Notes from the moderator",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.ReportReason": {
            "type": "string",
            "enum": [
                "SPAM",
                "HARASSMENT",
                "NUDITY",
                "HATE_SPEECH",
                "OTHER"
            ],
            "x-enum-varnames": [
                "ReportReasonSpam",
                "ReportReasonHarassment",
                "ReportReasonNudity",
                "ReportReasonHateSpeech",
                "ReportReasonOther"
            ]
        },
        "model.ReportedEntityType": {
            "type": "string",
            "enum": [
                "BOARD",
                "COMMENT",
                "USER",
                "GAME"
            ],
            "x-enum-comments": {
                "ReportedEntityTypeComment": "For both BoardComment and GameComment"
            },
            "x-enum-descriptions": [
                "",
                "For both BoardComment and GameComment",
                "",
                ""
            ],
            "x-enum-varnames": [
                "ReportedEntityTypeBoard",
                "ReportedEntityTypeComment",
                "ReportedEntityTypeUser",
                "ReportedEntityTypeGame"
            ]
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.Space": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "boardCollections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BoardCollection"
                    }
                },
                "boardComments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BoardComment"
                    }
                },
                "boardVotes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BoardVote"
                    }
                },
                "boards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Board"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                },
                "following": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                },
                "gameComments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GameComment"
                    }
                },
                "gameVotes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GameVote"
                    }
                },
                "games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Notification"
                    }
                },
                "reports": {
                    "description": "Reports made by this user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Report"
                    }
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                },
                "roleId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}